Split up AST into different files
