+++ source/Nevermore/AST/ISelectSource.cs
@@ -1,0 +1,12 @@
+﻿namespace Nevermore.AST
+{
+    public interface ISelectSource
+    {
+        string GenerateSql();
+    }
+
+    public interface IAliasedSelectSource : ISelectSource
+    {
+        string Alias { get; }
+    }
+}
