+++ source/Nevermore/AST/Over.cs
@@ -1,0 +1,16 @@
+﻿namespace Nevermore.AST
+{
+    public class Over
+    {
+        readonly OrderBy orderBy;
+        readonly PartitionBy partitionBy; // Can be null
+
+        public Over(OrderBy orderBy, PartitionBy partitionBy)
+        {
+            this.orderBy = orderBy;
+            this.partitionBy = partitionBy;
+        }
+
+        public string GenerateSql() => $"OVER ({(partitionBy == null ? string.Empty : $"{partitionBy.GenerateSql()} ")}{orderBy.GenerateSql()})";
+    }
+}
